introducing LAXATIVES
removed test objects from the file, no longer needed.
commented out section of the pigeon fly script to allow the poop to spawn. using the instantiate line of code in that script, you do not require the 3rd argument to spawn successfully.

TODO: add in a poo splat png image, and make it appear and fade over a short time after poop collides with player.